Move session check to its own file
