+++ gatsby-browser.js
@@ -4,10 +4,6 @@
  * See: https://www.gatsbyjs.org/docs/browser-apis/
  */
 
-import React, { useEffect, useState } from "react"
-import PropTypes from "prop-types"
-import { silentAuth } from "./src/utils/auth"
-
 import "./src/css/fonts.css"
 import "./src/css/base.css"
 import "./src/css/grid.css"
@@ -22,24 +18,6 @@ import "./src/css/sidebar.css"
 import "./src/css/conduct.css"
 import "./src/css/footer.css"
 
-const SessionCheck = ({ children }) => {
-	const [loading, setLoading] = useState(true)
-	useEffect(() => {
-		silentAuth(function() {
-			setLoading(false)
-		})
-	}, [])
-	return loading ? null : <React.Fragment>{children}</React.Fragment>
-}
-
-SessionCheck.propTypes = {
-	children: PropTypes.node
-}
-
-export const wrapRootElement = ({ element }) => (
-	<SessionCheck>{element}</SessionCheck>
-)
-
-wrapRootElement.propTypes = {
-	element: PropTypes.object.isRequired
-}
+import SessionCheck from "./session-check"
+const wrapRootElement = SessionCheck
+export default { wrapRootElement }
